move component to a isolated file
